Removing __init__.py and tescases, as it makes the repo structure more python oriented, than rather genric solution platform
-Removed the tescase folder
-Removed all references to __init__.py
-Moved tescases inside main, so others canstill beneift from the run to see if there results matches the expected.
