--- Programming/Python/AddTwoNumbers/solution/__init__.py
@@ -1,3 +1,0 @@
-from .addTwoNumbers import get_input_reverse_sum
-from .addTwoNumbers import get_sum_of_two_numbers
-from .addTwoNumbers import get_sum_of_two_numbers_39
